[refactor] Renaming ao.sparsity to ao.prune
`Sparsity` as a term doesn't reflect the tools that are developed by the AO.
The `torch/ao/sparsity` also has utilities for structured pruning, which
internally we always referred to as just "pruning". To avoid any confusion,
we renamed `Sparsity` to `Prune`. We will not be introducing the backwards
compatibility, as so far this toolset was kept under silent development.

This change will reflect the changes in the documentation as well.

**TODO:**
- [] Change the tutorials
- [] Confirm no bc-breakages
- [] Reflect the changes in the trackers and RFC docs
