+++ benchmarks/operator_benchmark/pt/ao_sparsifier_test.py
@@ -3,7 +3,7 @@ import operator_benchmark as op_bench
 import torch
 from torch import nn
 
-from torch.ao import sparsity
+from torch.ao import pruning
 
 
 """Microbenchmarks for sparsifier."""
@@ -33,7 +33,7 @@ class WeightNormSparsifierBenchmark(op_bench.TorchBenchmarkBase):
         model.register_buffer("weight", weight)
 
         sparse_config = [{"tensor_fqn": "weight"}]
-        self.sparsifier = sparsity.WeightNormSparsifier(
+        self.sparsifier = pruning.WeightNormSparsifier(
             sparsity_level=SL,
             sparse_block_shape=SBS,
             zeros_per_block=ZPB,
