+++ test/ao/sparsity/test_composability.py
@@ -7,10 +7,10 @@ import logging
 import torch
 import torch.ao.quantization as tq
 from torch import nn
-from torch.ao import sparsity
+from torch.ao import pruning
 from torch.testing._internal.common_utils import TestCase
 from torch.ao.quantization.quantize_fx import prepare_fx, convert_fx, convert_to_reference_fx, prepare_qat_fx
-from torch.ao.sparsity import fqn_to_module
+from torch.ao.pruning import fqn_to_module
 
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
@@ -37,7 +37,7 @@ def _get_model_and_sparsifier_and_sparse_config(qconfig=None):
         model[4].qconfig = qconfig
         model[5].qconfig = qconfig
 
-    sparsifier = sparsity.WeightNormSparsifier(**sparse_defaults)
+    sparsifier = pruning.WeightNormSparsifier(**sparse_defaults)
 
     sparse_config = [
         {
