+++ test/ao/sparsity/test_kernels.py
@@ -13,7 +13,7 @@ import torch
 import torch.ao.quantization as tq
 
 from torch import nn
-from torch.ao.sparsity.sparsifier.utils import fqn_to_module
+from torch.ao.pruning.sparsifier.utils import fqn_to_module
 
 from torch.testing._internal.common_utils import TestCase, skipIfTorchDynamo
 from torch.testing._internal.common_quantized import (
