+++ test/ao/sparsity/test_pruner.py
@@ -7,7 +7,7 @@ import logging
 
 import torch
 from torch import nn
-from torch.ao.sparsity._experimental.pruner import BasePruner, PruningParametrization, ZeroesParametrization
+from torch.ao.pruning._experimental.pruner import BasePruner, PruningParametrization, ZeroesParametrization
 from torch.nn.utils import parametrize
 
 from torch.testing._internal.common_utils import TestCase, skipIfTorchDynamo
