+++ torch/ao/pruning/_experimental/data_scheduler/base_data_scheduler.py
@@ -30,7 +30,7 @@ class BaseDataScheduler(object):
     def __init__(self, data_sparsifier, schedule_param: str, last_epoch=-1, verbose=False):
         # Attach sparsifier
         if not isinstance(data_sparsifier, BaseDataSparsifier):
-            raise TypeError('{} is not an instance of torch.ao.sparsity.BaseDataSparsifier'.format(
+            raise TypeError('{} is not an instance of torch.ao.pruning.BaseDataSparsifier'.format(
                 type(data_sparsifier).__name__))
         self.data_sparsifier = data_sparsifier
         self.schedule_param = schedule_param
