+++ torch/ao/pruning/_experimental/data_sparsifier/benchmarks/evaluate_model_metrics.py
