+++ torch/ao/pruning/_experimental/data_sparsifier/benchmarks/images/disk_savings.png
