+++ torch/ao/pruning/_experimental/data_sparsifier/benchmarks/images/forward_time.png
