+++ torch/ao/pruning/_experimental/pruner/images/prune_1.png
