+++ torch/ao/pruning/_experimental/pruner/images/prune_2.png
