+++ torch/ao/pruning/_experimental/pruner/images/prune_3.png
