+++ torch/ao/pruning/_experimental/pruner/images/prune_4.png
